+++ toxav/toxav.h
@@ -117,7 +117,7 @@ typedef enum {
  * @return ToxAv*
  * @retval NULL On error.
  */
-ToxAv *toxav_new(Tox *messenger, void *useragent, const char *ua_name, uint16_t video_width, uint16_t video_height);
+ToxAv *toxav_new(Tox *messenger, void *useragent, const char *ua_name, size_t video_width, size_t video_height);
 
 /**
  * @brief Remove A/V session.
